feat(output): add variables list extraction to v2
