+++ antarest/output/output_service.py
@@ -52,7 +52,6 @@ from antarest.output.aggregator_management import (
     LINK_COL,
 )
 from antarest.output.output_model import (
-    OutputVariables,
     OutputVariablesInformation,
     OutputVariablesList,
     OutputVariablesViewResponse,
@@ -74,15 +73,14 @@ from antarest.output.utils import (
     add_time_index_to_dataframe,
     split_concatenated_columns_from_dataframe,
 )
-from antarest.output.variables_management import (
+from antarest.output.variable_view.db import create_output_view_db_model, get_output_view_inside_db
+from antarest.output.variable_view.matrix_usage_provider import OutputVariablesMatrixUsageProvider
+from antarest.output.variable_view.model import (
     OutputItemId,
     check_output_variable_exists,
-    create_output_view_db_model,
     get_ids_for_aggregation,
-    get_output_view_inside_db,
     get_query_file,
 )
-from antarest.output.variables_matrix_usage_provider import OutputVariablesMatrixUsageProvider
 from antarest.study.model import (
     MatrixAggregationResultDTO,
     MatrixFrequency,
@@ -769,26 +767,11 @@ class OutputService:
     def get_output_variables_list(self, study_id: str, output_id: str) -> OutputVariablesList:
         """
         Returns the list of variables concerning a given output.
-        First, try to fetch the given data inside DB.
-        If present, return the data.
-        If not, parse the output headers to build the object. Before returning it, save it inside DB for next calls.
         """
         self._studies_repository.assert_permission(study_id, StudyPermissionType.READ)
 
-        output_variables: OutputVariables | None = db.session.get(OutputVariables, (study_id, output_id))
-        if output_variables:
-            return output_variables.to_model()
-
-        # Fetches the data from stored output
-        model = self._find_output_storage(study_id, output_id).extract_variables_list(study_id, output_id)
-
-        # Save the model inside DB for next calls
-        db_model = OutputVariables.from_model(study_id, output_id, model)
-        db.session.add(db_model)
-        db.session.commit()
-
-        # Returns it
-        return model
+        storage = self._find_output_storage(study_id, output_id)
+        return storage.get_variables_list(study_id, output_id)
 
     def get_output_variables_information(self, study_id: str, output_id: str) -> OutputVariablesInformation:
         """
