+++ antarest/output/storage/file/file_output_storage.py
@@ -42,7 +42,9 @@ from antarest.core.utils.utils import StopWatch
 from antarest.launcher.adapters.abstractlauncher import SimulationLogs
 from antarest.launcher.model import LogType
 from antarest.output.aggregator_management import AggregatorManager
+from antarest.output.filestudy.file_output_utils import extract_variables_list
 from antarest.output.output_model import OutputVariablesList
+from antarest.output.storage.file.repository import FileOutputRepository
 from antarest.output.storage.output_storage import (
     IOutputStorage,
     OutputDetails,
@@ -50,7 +52,6 @@ from antarest.output.storage.output_storage import (
     OutputStorageType,
 )
 from antarest.output.utils import QueryFileType
-from antarest.output.variables_management import extract_variables_list
 from antarest.study.model import (
     DEFAULT_WORKSPACE_NAME,
     MatrixFrequency,
@@ -198,10 +199,17 @@ class InStudyFileOutputStorage(IOutputStorage):
     Implementation based on outputs stored in antares-solver file format, inside a study.
     """
 
-    def __init__(self, outputs_provider: IFileOutputsProvider, cache: ICache, remote_executor: IRemoteExecutor) -> None:
+    def __init__(
+        self,
+        outputs_provider: IFileOutputsProvider,
+        cache: ICache,
+        remote_executor: IRemoteExecutor,
+        repository: FileOutputRepository,
+    ) -> None:
         self._outputs_provider = outputs_provider
         self._cache = cache
         self._remote_executor = remote_executor
+        self._repository = repository
 
     @override
     @property
@@ -484,9 +492,21 @@ class InStudyFileOutputStorage(IOutputStorage):
         return aggregator_manager.aggregate_output_data()
 
     @override
-    def extract_variables_list(self, study_id: str, output_id: str) -> OutputVariablesList:
+    def get_variables_list(self, study_id: str, output_id: str) -> OutputVariablesList:
         study_outputs = self._outputs_provider.get_outputs(study_id)
-        return extract_variables_list(_output_path(study_outputs.outputs_path, output_id))
+
+        output_variables = self._repository.get_output_variables_list(study_id, output_id)
+        if output_variables:
+            return output_variables
+
+        # Fetches the data from stored output
+        output_dir = _output_path(study_outputs.outputs_path, output_id)
+        model = extract_variables_list(output_dir)
+
+        # Save the model inside DB for next calls
+        self._repository.save_output_variables_list(study_id, output_id, model)
+
+        return model
 
     @override
     def get_logs(self, study_id: str, output_id: str, log_type: LogType) -> str:
