+++ antarest/study/dao/database/sql_utils.py
@@ -16,7 +16,7 @@ Collection of SQL related utilities
 
 from typing import Any
 
-from sqlalchemy import Column, Table, and_, insert, select, tuple_, update
+from sqlalchemy import Column, Table
 from sqlalchemy.dialects.postgresql import insert as pg_insert
 from sqlalchemy.dialects.postgresql.dml import Insert as PgInsert
 from sqlalchemy.dialects.sqlite import insert as sqlite_insert
@@ -36,57 +36,6 @@ def _update_columns(table: Table) -> list[Column[Any]]:
     Returns columns that are not part of the primary key, which will all be assumed updated.
     """
     return [column for column in table.columns.values() if not column.primary_key]
-
-
-def generic_upsert_multiple(
-    session: Session,
-    table: Table,
-    values: list[dict[str, Any]],
-) -> None:
-    """
-    Upsert implementation which does not rely on a specific dialect.
-
-    Note that it could fail on concurrency issues, since it relies on several sequential
-    requests.
-
-    Notes:
-        Assumes:
-         - all columns are updated
-         - value dictionaries all have the correct columns: we don't check for this
-    """
-    if not values:
-        return
-
-    prototype_value = values[0]
-
-    key_columns = _key_columns(table)
-    update_columns = [c for c in _update_columns(table) if c in prototype_value]
-
-    # First a select to identify existing rows
-    def _key(vals: dict[str, Any]) -> tuple[Any, ...]:
-        return tuple(vals[column.name] for column in key_columns)
-
-    key_tuples = [_key(row) for row in values]
-    existing_stmt = select(*key_columns).where(tuple_(*key_columns).in_(key_tuples))
-    existing_rows = session.execute(existing_stmt).fetchall()
-    existing_keys = {tuple(getattr(row, column.name) for column in key_columns) for row in existing_rows}
-
-    rows_to_insert: list[dict[str, Any]] = []
-    rows_to_update: list[dict[str, Any]] = []
-    for row in values:
-        if _key(row) in existing_keys:
-            rows_to_update.append(row)
-        else:
-            rows_to_insert.append(row)
-
-    if rows_to_insert:
-        session.execute(insert(table).values(rows_to_insert))
-
-    # 1 query per update unfortunately. Might be possible to optimize, but not worth it for us now.
-    for row in rows_to_update:
-        filters = [column == row[column.name] for column in key_columns]
-        stmt_update = update(table).where(and_(*filters)).values({col: row[col.name] for col in update_columns})
-        session.execute(stmt_update)
 
 
 def upsert_one(
@@ -142,4 +91,4 @@ def upsert_multiple(
         )
         session.execute(stmt)
     else:
-        generic_upsert_multiple(session, table, values)
+        raise NotImplementedError(f"Dialect {dialect} not supported")
