+++ tests/integration/maintenance_blueprint/test_gc_tasks.py
@@ -51,7 +51,7 @@ class TestTasksGCIntegration:
             assert len(task_list) == 4
 
             # deleting tasks
-            with patch("antarest.core.utils.utils.current_time", return_value=datetime(2026, 2, 20, 0, 0, 0)):
+            with patch("antarest.core.tasks.repository.current_time", return_value=datetime(2026, 2, 20, 0, 0, 0)):
                 task_result = clean_tasks(task_service=task_service, dry_run=False, task_retention_duration=17)
 
             expected_task = TaskDTO(
