+++ tests/output/test_output_service.py
@@ -24,7 +24,11 @@ from antarest.core.tasks.model import TaskDTO, TaskResult, TaskStatus, TaskType
 from antarest.core.tasks.service import ITaskService
 from antarest.core.utils.utils import current_time
 from antarest.output.output_service import IStudyMetadataProvider, OutputService, StudyMetadata
-from antarest.output.storage.file_output_storage import FileStudyOutputs, IFileOutputsProvider, InStudyFileOutputStorage
+from antarest.output.storage.file.file_output_storage import (
+    FileStudyOutputs,
+    IFileOutputsProvider,
+    InStudyFileOutputStorage,
+)
 from antarest.output.storage.output_storage import IOutputStorage, OutputStorageType
 from antarest.study.model import (
     RawStudy,
@@ -106,7 +110,10 @@ def test_unarchive_output_for_other_workspace_is_executed_on_remote(
     remote_executor: IRemoteExecutor = Mock(spec=IRemoteExecutor)
     file_outputs_provider = _file_outputs_provider(study_mock)
     output_storage = InStudyFileOutputStorage(
-        cache=cache_mock, outputs_provider=file_outputs_provider, remote_executor=remote_executor
+        cache=cache_mock,
+        outputs_provider=file_outputs_provider,
+        remote_executor=remote_executor,
+        repository=Mock(),
     )
 
     studies_metadata_repository = _studies_repository(study_mock)
@@ -161,7 +168,7 @@ def test_archive_output_locks(tmp_path: Path, command_context: CommandContext) -
     task_service = Mock(spec=ITaskService)
     file_outputs_provider = _file_outputs_provider(study_mock)
     output_storage = InStudyFileOutputStorage(
-        cache=cache_mock, outputs_provider=file_outputs_provider, remote_executor=Mock()
+        cache=cache_mock, outputs_provider=file_outputs_provider, remote_executor=Mock(), repository=Mock()
     )
 
     studies_metadata_repository = _studies_repository(study_mock)
