+++ tests/output/test_variables_manager.py
@@ -13,7 +13,7 @@ import pytest
 
 from antarest.core.exceptions import OutputVariablesViewError
 from antarest.output.output_model import OutputVariablesList
-from antarest.output.variables_management import (
+from antarest.output.variable_view.model import (
     AreaOutputId,
     LinkOutputId,
     OutputItemId,
